--- ansible/ioc_module/global_inventory.ini
@@ -1,58 +1,0 @@
-; Facilities Global Inventory *used by all deployment playbooks
-
-;   TODO: Need to figure out the password (ssh key? ansible-vault-pass?)
-;   TODO: fill in rest of vars for the other facilities (look at .cram_user.facilities.cfg)
-;   TODO: For now we are testing on s3df, but we need a proxy jump in real deal
-
-[S3DF]
-; we can assume deployments will be done from dev server for now
-localhost ansible_connection=local
-; s3dflogin.slac.stanford.edu ansible_user=adbuild 
-; adbuild@s3df.slac.stanford.edu 
-
-[LCLS]
-lcls-srv01 ansible_user=adbuild
-; softegr@lcls-srv01
-
-[FACET]
-facet-srv01 ansible_user=adbuild
-; fsoftegr@facet-srv01
-
-[TestFac]
-testfac-srv01 ansible_user=adbuild
-; tfprod@testfac-srv01
-
-[S3DF:vars]
-base_path=/sdf/group/ad/eed/lcls
-ioc_release_folder=/sdf/group/ad/eed/lcls/epics/iocTop
-ioc_link_folder=/sdf/group/ad/eed/lcls/epics/iocCommon
-ioc_data_folder=/sdf/group/ad/eed/lcls/epics/iocData
-hla_release_folder=/sdf/group/ad/eed/lcls/physics/release
-hla_link_folder=/sdf/group/ad/eed/lcls/physics
-tools_release_folder=/sdf/group/ad/eed/lcls/tools/script/release
-tools_link_folder=/sdf/group/ad/eed/lcls/tools/script
-matlab_release_folder=/sdf/group/ad/eed/lcls/tools/matlab/toolbox/release
-matlab_link_folder=/sdf/group/ad/eed/lcls/tools/matlab/toolbox
-pydm_release_folder=/sdf/group/ad/eed/lcls/tools/pydm/display/release
-pydm_link_folder=/sdf/group/ad/eed/lcls/tools/pydm/display
-
-[LCLS:vars]
-ioc_release_folder=/usr/local/lcls/epics/iocTop
-ioc_link_folder=/usr/local/lcls/epics/iocCommon
-ioc_data_folder=/usr/local/lcls/epics/iocData
-hla_release_folder=/usr/local/lcls/physics/release
-hla_link_folder=/usr/local/lcls/physics
-tools_release_folder=/usr/local/lcls/tools/script/release
-tools_link_folder=/usr/local/lcls/tools/script
-matlab_release_folder=/usr/local/lcls/tools/matlab/toolbox/release
-matlab_link_folder=/usr/local/lcls/tools/matlab/toolbox
-pydm_release_folder=/usr/local/lcls/tools/pydm/display/release
-pydm_link_folder=/usr/local/lcls/tools/pydm/display
-; TODO: Proxy into mcc
-; ansible_ssh_common_args='-o ProxyCommand="ssh -W %h:%p proxy.server.address"'
-
-[FACET:vars]
-epics_path=/usr/local/lcls/epics
-
-[TestFac:vars]
-epics_path=/usr/local/lcls/epics
