+++ bs_cli/src/adbs_cli/component.py
@@ -16,9 +16,6 @@ class Component(object):
         except:
             return None
         
-    def git_get_top_dir(self):
-        return self.git_repo.git.rev_parse("--show-toplevel")
-        
     def git_clone(self, url: str):
         # Clone repo from github url
         self.git_repo.git.clone(url)
@@ -32,7 +29,7 @@ class Component(object):
 
     def git_commit(self, branch_name: str):
         # Commit empty
-        commit_msg = 'initial commit for #' + branch_name
+        commit_msg = 'initial commit for ' + branch_name
         self.git_repo.git.commit("--allow-empty", '-m', commit_msg)
 
     def git_push(self, branch_name: str):
