--- other/sample_rpm_ioc_spec
@@ -1,55 +1,0 @@
-Name:           test-ioc
-Version:        1.0.0
-Release:        1%{?dist}
-Summary:        test-ioc built for buildsystem testing
-BuildArch:      x86_64
-
-License:        GPL
-Source0:        %{name}-%{version}.tar.gz
-# Define the _prefix macro with a default value
-Prefix: /iocTop
-
-# Requires:       
-# TODO: Need to figure out if we want to specify dependencies here like all of them?
-# Or just copy over the dependencies from the manifest CONFIG.yaml
-
-%description
-A demo RPM build
-
-%prep
-%setup -q
-
-# %build
-# No build steps for this - can assume the tarball is already built
-
-%install
-# Install refers to installing the tarball on BUILDROOT/ of the rpm package
-# Create the test-ioc directory in the buildroot
-mkdir -p %{buildroot}/iocTop/%{name}-%{version}/bin
-mkdir -p %{buildroot}/iocTop/%{name}-%{version}/db
-mkdir -p %{buildroot}/iocTop/%{name}-%{version}/dbd
-mkdir -p %{buildroot}/iocTop/%{name}-%{version}/iocBoot
-
-# Copy contents from the unpacked tarball into the test-ioc directory
-cp -a bin/* %{buildroot}/iocTop/%{name}-%{version}/bin
-cp -a db/* %{buildroot}/iocTop/%{name}-%{version}/db
-cp -a dbd/* %{buildroot}/iocTop/%{name}-%{version}/dbd
-cp -a iocBoot/* %{buildroot}/iocTop/%{name}-%{version}/iocBoot
-
-%clean
-# No clean steps for this - since there is no building step
-
-%files
-# Files refer to copying the files to desired directories on target system
-# The install location is dynamic and determined at installation using 'relocate'
-# ex: rpm -i test-ioc.rpm --relocate /iocTop=/sdf/scratch/ad/build/lcls/epics/iocTop
-# This is recursive, all files under this directory will be copied, and removed during uninstall
-/iocTop/%{name}-%{version}
-
-## Fixes
-# disable shebang mangling of python scripts
-%undefine __brp_mangle_shebangs
-
-%changelog
-* Fri Aug 23 2024 Patrick Nisperos <pnispero@slac.stanford.edu> - 1.0.0
-- First version being packaged
